Develop: функционал Этапа 1

+++ statistics/stat-dto/src/main/java/ru/practicum/dto/ViewStats.java
@@ -11,5 +11,5 @@ import lombok.*;
 public class ViewStats {
     private String app;
     private String uri;
-    private int hits;
+    private Long hits;
 }
